♻️ Migrate the unit tests of the hot loading prompt provider into the correct naming test module.
